+++ source_code/system_hydraulic/source_code/solvergpu/source_code/util.h
@@ -24,16 +24,12 @@ struct cursorCoords {
  */
 namespace Util
 {
-	// String conversions
-	std::string		secondsToTime(double);
-
-	// Resource handling
-	char*			getFileResource(const char*, const char*);
-	double			round(double, unsigned int);
-	std::string     to_string_exact(double);
-	std::string		get_error_str(int);
-
-	void			SwapEnd(float& var);
+	std::string		secondsToTime(double);						// Get time as a string
+	char*			getFileResource(const char*, const char*);	// Resource handling
+	double			round(double, unsigned int);				// Round a number to set decimal places
+	std::string     to_string_exact(double);					// Convert double to string (useful for extra small or extra large doubles like 1e-10)
+	std::string		get_error_str(int);							// Returns a string of an OpenCL error
+	void			SwapEnd(double& var);						// In-place byte order swap for a double-precision floating-point number.
 
 }
